+++ src/module/feature/loops/models/detected-clan.model.ts
@@ -4,7 +4,6 @@ import { Api } from '../../../shared/decorators/injector/api-injector.decorator'
 import { LoggerInjector } from '../../../shared/decorators/injector/logger-injector.decorator';
 import { Table } from '../../../shared/decorators/injector/table-injector.decorator';
 import type { LeavingPlayersTable } from '../../../shared/tables/complexe-table/leaving-players/leaving-players.table';
-import type { Clan } from '../../../shared/tables/complexe-table/watch-clans/models/watch-clans.type';
 import type { WatchClansTable } from '../../../shared/tables/complexe-table/watch-clans/watch-clans.table';
 import type { PotentialClansTable } from '../../../shared/tables/simple-table/potential-clans.table';
 import type { Logger } from '../../../shared/utils/logger';
@@ -12,11 +11,15 @@ import { ChannelsTable } from '../../../shared/tables/complexe-table/channels/ch
 import { Client, Colors, EmbedBuilder, TextChannel } from 'discord.js';
 import { fetchChannelFromClient } from '../../../shared/utils/user.util';
 import { transformToCode } from '../../../shared/utils/string.util';
+import { WargamingApi } from '../../../shared/apis/wargaming/wargaming.api';
+import { wording } from '../../../shared/utils/config';
+import { WargamingClanInfo } from '../../../shared/apis/wargaming/models/wargaming.type';
 
 @LoggerInjector
 export class DetectedClanModel {
     //region INJECTABLE
     @Api('Wot') private readonly wotApi: WotApi;
+    @Api('Wargaming') private readonly wargamingApi: WargamingApi;
     @Table('WatchClans') private readonly watchClans: WatchClansTable;
     @Table('LeavingPlayers') private readonly leavingPlayers: LeavingPlayersTable;
     @Table('PotentialClans') private readonly potentialClans: PotentialClansTable;
@@ -54,11 +57,10 @@ export class DetectedClanModel {
             }
 
             const clanId = datum.clan_id;
-            const clans: Clan[] = await this.watchClans.selectClan(String(clanId));
 
-            if (clanId !== null && clanId !== 500312605 && clans.length === 0 && !(await this.potentialClans.clanExist(clanId))) {
+            if (await this.isPotentialClan(clanId)) {
                 this.logger.debug('Clan found from leaving player : {}', clanId);
-                await this.potentialClans.addClan(clanId);
+                await this.potentialClans.addClan(clanId!);
                 this.numberOfClansDetected++;
             }
 
@@ -69,6 +71,23 @@ export class DetectedClanModel {
     }
 
     /**
+     * Check if the clan of the player is a potential clan to watch
+     *
+     * @param {number | null} clanId - The clan id of the player, if null, the player doesn't has clan
+     *
+     * @return {Promise<boolean>} - True if the clan is a potential clan, false otherwise
+     */
+    private async isPotentialClan(clanId: number | null): Promise<boolean> {
+        return (
+            clanId !== null &&
+            clanId !== 500312605 &&
+            (await this.watchClans.selectClan(String(clanId))).length === 0 &&
+            !(await this.potentialClans.clanExist(clanId)) &&
+            (await this.isFrenchClan(clanId))
+        );
+    }
+
+    /**
      * Send message to announce that the detection of clan is finished
      */
     public async sendMessageClanDetectedReady(): Promise<void> {
@@ -76,14 +95,31 @@ export class DetectedClanModel {
             embeds: [
                 new EmbedBuilder()
                     .setColor(Colors.DarkGold)
-                    .setTitle('La détection de clan vient de se terminer !')
+                    .setTitle(wording('detected-clan.texts.embed-clan-detected.title'))
                     .setDescription(
-                        transformToCode(
-                            "Un total de {} clan a été détecté parmi l'ensemble des joueurs ayant quitté leur clan. Utilisez la commande `/detected-clan` pour voir les clans.",
-                            this.numberOfClansDetected
-                        )
+                        transformToCode(wording('detected-clan.texts.embed-clan-detected.description'), this.numberOfClansDetected)
                     ),
             ],
         });
+    }
+
+    /**
+     * Check if the clan of the leaving player is a french clan.
+     *
+     * @param {number} clanId - The clan id to check
+     *
+     * @return {Promise<boolean>} - True if is a french tank, false otherwise
+     */
+    private async isFrenchClan(clanId: number): Promise<boolean> {
+        const clanInfo: WargamingClanInfo = await this.wargamingApi.clanInfo(clanId);
+
+        if (!clanInfo) {
+            this.logger.error(wording('detected-clan.errors.failed-clan-info'), clanId);
+            return true;
+        }
+
+        return !!clanInfo.clanview.profiles
+            .find(({ type }): boolean => type === 'clan')
+            ?.languages_list?.find((language: string): boolean => language === 'fr');
     }
 }